% Generated by roxygen2 (4.0.1): do not edit by hand
\name{set_func_beta_dV}
\alias{set_func_beta_dV}
\title{Set the evapotranspiration reduction function}
\usage{
set_func_beta_dV(newfunc)
}
\arguments{
\item{newfunc}{a function which computes evapotranspiration reduction \code{beta}
(=ETact/ETpot) [unitless] from storage deficit \code{dV} [mm],
with arguments \code{x} (here storage deficit).}
}
\value{
a function to be used as evapotranspiration reduction function.
}
\description{
Changes the default evapotranspiration reduction function \code{func_beta_dV}
}
\examples{
x=1
}

