% Generated by roxygen2 (4.0.1): do not edit by hand
\name{show_func_Q_hS}
\alias{show_func_Q_hS}
\title{Show current stage-discharge relation}
\usage{
show_func_Q_hS()
}
\value{
the function to be used as stage-discharge relation.
}
\description{
Shows which stage-discharge relation \code{func_Q_hS} is used.
}
\examples{
x=1
}

