% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{WALRUS}
\alias{WALRUS}
\alias{WALRUS-package}
\title{The Wageningen Lowland Runoff Simulator}
\description{
WALRUS is a rainfall-runoff model for catchments with shallow groundwater.
This package provides all model code and examples to run WALRUS.
}
\details{
A user manual is provided on www.GitHub.com/ClaudiaBrauer/WALRUS,
containing explainions of the model structure, model variables, model parameters and
what these represent. It also gives detailed tips on how to get WALRUS
running for your own catchment.
}
\references{
C.C. Brauer, A.J. Teuling, P.J.J.F. Torfs, R. Uijlenhoet (2014a):
The Wageningen Lowland Runoff Simulator (WALRUS):
a lumped rainfall-runoff model for catchments with shallow groundwater,
Geoscientific Model Development, 7, 2313-2332,
\url{http://www.geosci-model-dev.net/7/2313/2014/gmd-7-2313-2014.pdf}

C.C. Brauer, P.J.J.F. Torfs, A.J. Teuling, R. Uijlenhoet (2014b):
The Wageningen Lowland Runoff Simulator (WALRUS):
application to the Hupsel Brook catchment and Cabauw polder,
Hydrology and  Earth System Sciences, 18, 4007-4028,
\url{www.hydrol-earth-syst-sci.net/18/4007/2014/hess-18-4007-2014.pdf}
}

