% Generated by roxygen2 (4.0.1): do not edit by hand
\name{WALRUS_date_conversion}
\alias{WALRUS_date_conversion}
\title{Converts date format.}
\usage{
WALRUS_date_conversion(d)
}
\arguments{
\item{d}{date/time combination in date or POSIX-format}
}
\value{
number of seconds since 1970-01-01
}
\description{
Converts YYYmmddHH to seconds since 1970-01-01.
}
\examples{
x=1
}

