% Generated by roxygen2 (4.0.1): do not edit by hand
\name{WALRUS_loop}
\alias{WALRUS_loop}
\title{Runs over all time steps and calls \code{WALRUS_step}}
\usage{
WALRUS_loop(pars)
}
\arguments{
\item{pars}{the parameter set used for the run.}
}
\value{
a data frame with the model output for all output time steps.
}
\description{
One of the main WALRUS-functions. Computes initial values and provides the for-loop with flexible time step.
}
\examples{
x=1
}

